Add files via upload
So, this is a somewhat updated version of my score keeper. This is something I really want to build in an advanced way. My friends and I always play certain games, and it would be really cool and useful to have a score keeper that would be able to save players, and scores that would map to those players. However, I cannot for the life of me figure out how to save user input in Python, so hopefully, the course I'm currently in will teach me ho, please . I'm thinking I probably need to connect SQL or SQLite in some way to properly hold a database of user and scores, but that is above me for now. If anybody sees this, and has any recommendations as to how to accomplish the goal I want to accomplish, please, lmk. Other than that, thank you so much for taking the time to view my code!